+++ fabric/src/main/java/com/lx862/jcm/mixin/modded/mtr/RenderEyeCandyMixin.java
@@ -55,7 +55,7 @@ public class RenderEyeCandyMixin {
         storedMatrixTransformations.add((graphicsHolderNew) -> {
             graphicsHolderNew.translate(blockEntity.getTranslateX(), blockEntity.getTranslateY(), blockEntity.getTranslateZ());
             graphicsHolderNew.rotateYDegrees(180 - facing.asRotation());
-            graphicsHolderNew.rotateXRadians(blockEntity.getRotateX() + (float) Math.PI);
+            graphicsHolderNew.rotateXRadians(blockEntity.getRotateX());
             graphicsHolderNew.rotateYRadians(blockEntity.getRotateY());
             graphicsHolderNew.rotateZRadians(blockEntity.getRotateZ());
         });