+++ fabric/src/main/java/com/lx862/jcm/mod/scripting/jcm/JCMScripting.java
@@ -3,10 +3,11 @@ package com.lx862.jcm.mod.scripting.jcm;
 import com.lx862.jcm.mod.Constants;
 import com.lx862.jcm.mod.JCMClient;
 import com.lx862.jcm.mod.render.gui.ScriptDebugOverlay;
+import com.lx862.jcm.mod.scripting.jcm.pids.PIDSModelDrawCall;
 import com.lx862.jcm.mod.scripting.jcm.pids.TextWrapper;
 import com.lx862.jcm.mod.scripting.jcm.pids.TextureWrapper;
 import com.lx862.jcm.mod.scripting.mtr.MTRScripting;
-import com.lx862.jcm.mod.scripting.mtr.render.ModelDrawCall;
+import com.lx862.jcm.mod.scripting.mtr.render.DisplayHelperCompat;
 import com.lx862.jcm.mod.scripting.mtr.render.QuadDrawCall;
 import com.lx862.jcm.mod.scripting.mtr.util.ModelManager;
 import com.lx862.jcm.mod.scripting.mtr.util.TextUtil;
@@ -35,13 +36,14 @@ public class JCMScripting {
         scriptManager.getClassShutter().allowClass(ClassRule.parse("com.lx862.jcm.mod.scripting.mtr.*"));
 
         scriptManager.onParseScript((contextName, context, scriptable) -> {
-            scriptable.put("ModelDrawCall", scriptable, new NativeJavaClass(scriptable, ModelDrawCall.class));
             scriptable.put("QuadDrawCall", scriptable, new NativeJavaClass(scriptable, QuadDrawCall.class));
             scriptable.put("MTRClientData", scriptable, new NativeJavaClass(scriptable, MinecraftClientData.class));
             scriptable.put("TextUtil", scriptable, new NativeJavaClass(scriptable, TextUtil.class));
+            scriptable.put("DisplayHelperCompat", scriptable, new NativeJavaClass(scriptable, DisplayHelperCompat.class));
             scriptable.put("ModelManager", scriptable, new NativeJavaClass(scriptable, ModelManager.class));
 
             if (contextName.equals("PIDS")) {
+                scriptable.put("ModelDrawCall", scriptable, new NativeJavaClass(scriptable, PIDSModelDrawCall.class));
                 scriptable.put("Text", scriptable, new NativeJavaClass(scriptable, TextWrapper.class));
                 scriptable.put("Texture", scriptable, new NativeJavaClass(scriptable, TextureWrapper.class));
             }