+++ fabric/src/main/java/com/lx862/jcm/mod/scripting/mtr/util/ScriptedModel.java
@@ -40,6 +40,7 @@ public class ScriptedModel implements StoredModelResourceBase {
             if(optimizedModelWrapper != null) {
                 MainRenderer.scheduleRender(QueuedRenderLayer.TEXT, (graphicsHolder, offset) -> {
                     storedMatrixTransformations.transform(graphicsHolder, offset);
+                    graphicsHolder.rotateXDegrees(180F);
                     CustomResourceLoader.OPTIMIZED_RENDERER_WRAPPER.queue(optimizedModelWrapper, graphicsHolder, light);
                     graphicsHolder.pop();
                 });
@@ -48,6 +49,7 @@ public class ScriptedModel implements StoredModelResourceBase {
             if(dynamicVehicleModel != null) {
                 MainRenderer.scheduleRender(QueuedRenderLayer.TEXT, (graphicsHolder, offset) -> {
                     storedMatrixTransformations.transform(graphicsHolder, offset);
+                    graphicsHolder.rotateXDegrees(180F);
                     dynamicVehicleModel.render(graphicsHolder, light, 0, 1, 1, 1, 1);
                     graphicsHolder.pop();
                 });