+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # JCM v2.1.1 for MTR 4.0.2 has been released!
 
 ## Fixes
+- Fix Weather icon in JSON PIDS Preset being slightly off vertically
 - **Scripting:**
 - - Fix Draw Calls offset potentially affecting position of another draw call
 - - Fix `MinecraftClient#localPlayer()` being a non-static method, now you can invoke them directly like all other static method.